requirements.txt
numpy==2.1.2
packaging==24.1
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tenacity==9.0.0
tzdata==2024.2
